Stop logging manually failed jobs
This fixes an issue where jobs with uncaught exceptions are logged
multiple times. These exceptions are already reported to telescope, so
we don't need to listen for it.

For manually failed jobs, this means that you will have to use the
report() helper before calling the fail() method on the job.